Internal/2021.3/staging

+++ Packages/com.unity.render-pipelines.universal/ShaderLibrary/UnityInput.hlsl
@@ -119,7 +119,8 @@ float4 unity_RenderingLayer;
 half4 unity_LightData;
 half4 unity_LightIndices[2];
 
-half4 unity_ProbesOcclusion;
+// Don't use half4 here as that makes shaders not SRP batcher compatible
+float4 unity_ProbesOcclusion;
 
 // Reflection Probe 0 block feature
 // HDR environment map decode instructions
